requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy>=2.0,<2.2
alembic>=1.15,<2.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.6,<3.0
PyJWT>=2.10,<3.0
bcrypt>=4.2,<5.0
httpx>=0.28,<1.0
email-validator>=2.2,<3.0